lawyer profile done except bar associations

+++ lib/screens/profile/widget/profile_card.dart
@@ -179,7 +179,7 @@ Widget clientReview({
   String? userName,
   String? date,
   String? comment,
-  int? star,
+  double? star,
   like,
   dislike,
   Function()? likeTap,
